Split InterfaceSampleBased to Interface and InterfaceVillas
